+++ packages/url-loader/src/types/analytics.ts
@@ -1,8 +1,8 @@
 import { z } from 'zod';
 import { IAnalyticsOptions } from '@cloudinary/url-gen/sdkAnalytics/interfaces/IAnalyticsOptions';
 
-// Analytics Options
+export interface CloudinaryAnalyticsOptions extends IAnalyticsOptions {}
 
-export const analyticsOptionsSchema: z.ZodType<IAnalyticsOptions> = z.any();
+export const analyticsOptionsSchema: z.ZodType<CloudinaryAnalyticsOptions> = z.any();
 
 export type AnalyticsOptions = z.TypeOf<typeof analyticsOptionsSchema> | false;