+++ packages/url-loader/src/types/config.ts
@@ -1,9 +1,8 @@
 import { z } from 'zod';
-
-// Cloudinary config object
-
 import ICloudinaryConfigurations from '@cloudinary/url-gen/config/interfaces/Config/ICloudinaryConfigurations';
 
-export const configOptionsSchema: z.ZodType<ICloudinaryConfigurations> = z.any();
+export interface CloudinaryConfigurationOptions extends ICloudinaryConfigurations {}
+
+export const configOptionsSchema: z.ZodType<CloudinaryConfigurationOptions> = z.any();
 
 export type ConfigOptions = z.TypeOf<typeof configOptionsSchema>;